removing data member
